🏗️ Issue #1175 キーワードパーサー最適化完了: 650行→6ファイル責任分離実現
## 実装詳細
- UnifiedKeywordParser分割: 360行→274行（目標150-200行に近接）
- 設定管理分離: keyword_config.py（88行）で初期化ロジック独立
- バリデーション分離: keyword_validation.py（94行）で検証処理独立
- 抽出処理分離: keyword_extractors.py（118行）で情報処理独立
- ユーティリティ最適化: keyword_utils.py 280行→177行

## 技術的成果
- 各ファイル500行以下実現（最大329行）
- 責任分離による保守性向上
- 100%キーワード機能維持・テスト通過
- 6層アーキテクチャによる拡張性確保
